keep single-unit helpers out of the shared MainWindow header
Ten entities used by one unit go back to that unit's anonymous namespace, which
is what made the ODR bug writable in the first place. exceptionMessage was
triplicated across DatasetWindow, SequenceController and the header; it moves to
QtErrorText.hpp and the copies go. Header is self-contained and 529 -> 259
lines. 116/116, Clang -Werror clean.
