+++ src/qt/QtErrorText.hpp
@@ -1,0 +1,27 @@
+#pragma once
+
+#include <QException>
+#include <QString>
+
+#include <exception>
+
+namespace amrvis::qt {
+
+// Qt Concurrent masks worker exceptions behind QUnhandledException, so the
+// underlying library error text must be unwrapped before it is shown.
+inline QString exceptionMessage(const std::exception& error)
+{
+    const auto* unhandled = dynamic_cast<const QUnhandledException*>(&error);
+    if (unhandled != nullptr && unhandled->exception()) {
+        try {
+            std::rethrow_exception(unhandled->exception());
+        } catch (const std::exception& inner) {
+            return QString::fromUtf8(inner.what());
+        } catch (...) {
+            return QStringLiteral("unknown non-std exception");
+        }
+    }
+    return QString::fromUtf8(error.what());
+}
+
+} // namespace amrvis::qt
